+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 2.0.0
+
+- Update to Silero VAD v5
+- Audio size **must** be 512 samples in `process_chunk`
+- Add `process_chunks` to handle longer audio
+
 ## 1.0.0
 
 - Initial version